added functionality to delete a session_engineer by sessionId and engineerId

+++ studiomanagerapi/views/session_engineer.py
@@ -14,11 +14,6 @@ class SessionEngineerView(ViewSet):
             Response --  single JSON serialized session engineer dictionary
         """
         session_engineer = SessionEngineer.objects.get(pk=pk)
-        """Filter by engineer and session"""
-        engineer = request.query_params.get('engineer_id', None)
-        session = request.query_params.get('session_id', None)
-        if engineer is not None and session is not None:
-                session_engineer = SessionEngineer.filter(engineer_id=engineer, session_id=session)
         serializer = SessionEngineerSerializer(session_engineer)
         return Response(serializer.data)
 
@@ -68,9 +63,18 @@ class SessionEngineerView(ViewSet):
         Returns:
             Response -- Empty body with 204 status code
         """
-        session_engineer = SessionEngineer.objects.get(pk=pk)
-        session_engineer.delete()
-        return Response(None, status=status.HTTP_204_NO_CONTENT)
+        engineer = Engineer.objects.get(pk=pk)
+        session = request.query_params.get('session_id', None)
+        if session is not None:
+            # Filter the SessionEngineer instances for the matching engineer and session
+            session_engineer = SessionEngineer.objects.get(engineer_id=engineer, session_id=session)
+
+            if session_engineer:
+                # Delete the first matching session_engineer instance
+                session_engineer.delete()
+                return Response(None, status=status.HTTP_204_NO_CONTENT)
+
+        return Response({"detail": "SessionEngineer not found"}, status=status.HTTP_404_NOT_FOUND)
 
 class SessionEngineerSerializer(serializers.ModelSerializer):
     class Meta:
